+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 ### Fixed
 
 - Fixed issue with using Sparv as a library. Sparv can now be invoked multiple times within the same process.
+- Fixed issue with non-breaking spaces in misc:text_headtail.
 
 ## [5.3.1] - 2025-08-20